Add files via upload
Day - 03

Progress - 5 Questions
11)Maximum Points You Can Obtain from Cards
12)Container With Most Water
13)3Sum
14)4Sum
15)Find All Duplicates in an Array

What did I learn today ?

Today I found 3sum and 4sum problem very challenging but later figure out that it just based on the 2 sum problem and an approach of which I was unaware of.
Thanks to Striver. Rest of the Questions were on sliding window , greedy technique.